Add ignored topics for handling public channels
Public channels can be very expensive to track in a large system due to the way that a single topic
is sharded to the same value all the time. This causes a tracker node to run hot.

Probabilistically the topic will always have a listener and so we don't need to track it. We just always
broadcast a message.